+++ tests/e2e/walletBudgetToRecurring.spec.ts
@@ -1,10 +1,10 @@
-import { transformBalance } from '@/shared/helpers';
+import { test, expect } from './fixtures/connected';
+import { getStorage } from './fixtures/helpers';
+import { getContinueWaitTime, setupPlayground } from './helpers/common';
+import { completeGrant, DEFAULT_CONTINUE_WAIT_MS } from './helpers/testWallet';
+import { sendOneTimePayment } from './pages/popup';
 import { addMonths } from 'date-fns';
-import { test, expect } from '../fixtures/connected';
-import { getStorage } from '../fixtures/helpers';
-import { getContinueWaitTime, setupPlayground } from '../helpers/common';
-import { completeGrant, DEFAULT_CONTINUE_WAIT_MS } from '../helpers/testWallet';
-import { sendOneTimePayment } from '../pages/popup';
+import { transformBalance } from '@/shared/helpers';
 
 test('edit wallet budget from one-time to recurring', async ({
   popup,
